simulacra/requirements.txt
# Core dependencies
openai>=1.0.0,<2.0.0
openai-agents>=0.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0,<3.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Development
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.0.0
